Initial commit of script and ffmpeg binary
The script needs a couple of modifications before we make it public: 

- It needs to define version of FFmpeg and libvpx to fix them at a known stable release, with the option to remove the fixed version numbers and simply 'get latest' as it does in this version.

- It needs to support architectures which use apt-get rather than yum install.

The binary was built on a Google Cloud Platform n4-standard-4 Centos 7 instance. It has been tested on Centos 7,6 and Ubuntu 16.04. Please let us know if you have success with it on other architectures.

The build did not work on smaller instances: there are parts of the build which are quite memory intensive and we think the smaller instances simply struggled with available memory. It took about 35 mins to run on the standard size machine.

Former-commit-id: 9d8b59a09f68f6140bd368429e7b0c54a34b49fc